requirements.txt
fastapi
uvicorn

pydantic
nltk
numpy
scipy

mistralai
python-dotenv

asyncpg
pgvector
